Data Pipeline/requirements.txt
# Core Dependencies
chromadb>=0.4.22
sentence-transformers>=2.3.1
langchain>=0.1.0
langchain-community>=0.0.16
langgraph>=0.0.26

# Text Processing
langchain-text-splitters>=0.0.1

# Utilities
pyyaml>=6.0
numpy>=1.26.0
tqdm>=4.66.0

# Optional: For better performance
torch>=2.1.0
transformers>=4.36.0
